Version 1.0
This is an extremely barebones fabric mod that displays the number of total minecraft days in the top left of the players screen.

To change the day, use /setday <dayNumber>

Yes it's sort of broken and technically "adds" days instead of setting them
will I fix it? yes

Can you toggle it off? no

Can non-ops use the command? yes.

its kind of a piece of shit but its MY piece of shit and I plan to update it and maybe make it something I can feel proud of in future updates.

Also I feel the need to come clean and admit that a lot of this code comes from chatGPT. GPT made a lot of weird bugs, which made me end up having to learn how to do everything myself anyway, so I think it all balances out.