This took me far too long to complete. Many nights were spent stubbornly banging my head against a wall. The code is super messy, but hey, it works... I think. I'm going to test this build out on my own world, and if I notice anything break I'll come back and update this.
Everything is configured with the '/daycounterdisplay' or '/dcd' command. This mod also creates /.minecraft/DayCounterDisplay/daycounterdisplay-config.txt to save your settings. You could potentially edit that file but be careful with it. If you mess it up, just delete it and a new one will generate next time you join a world.

If anyone actually uses this, I hope you enjoy it! I know it doesn't look like much, but it took me A LOT of effort.

If you know how to program and want to upgrade/overhaul this mod feel free. I think for the most part I am done with this project. It would be cool if you credited me for the idea, but honestly idrc.